优化注册登陆
