+++ 91wangcai/app/Http/Controllers/Home/IndexController.php
@@ -8,10 +8,9 @@
 namespace App\Http\Controllers\Home;
 
 use App\Http\Controllers\Controller;
-use App\Http\Models\user;
+use App\Http\Models\UserModel;
 use Gregwar\Captcha\CaptchaBuilder;
 use Illuminate\Support\Facades\Input;
-
 
 /**
  * Class IndexController
@@ -32,18 +31,16 @@ class IndexController extends Controller
      */
     public function login()
     {
-        session_start();
         $user_phone = Input::get('user_phone');
         $user_password = md5(Input::get('user_password'));
         $code = Input::get('code');
-        $user = new user();
-        $phone = $user->phone($user_phone);
-        $login = $user->login($user_phone,$user_password);
-        if($_SESSION['milkcaptcha'] != $code) {
-            return json_encode(0);
+        if(session('milkcaptcha') != $code) {
+            return 0;
         }
+        $user = new UserModel();
+        $phone = $user->phone($user_phone);
         if(!$phone) {
-            return json_encode(1);
+            return 1;
         }
         if($phone) {
             $res = object_to_arrays($phone);
@@ -59,10 +56,11 @@ class IndexController extends Controller
                 return ['state'=>2,'text'=>$user_error['user_error']];
             }
         }
+        $login = $user->login($user_phone,$user_password);
         if($login) {
             $data = object_to_arrays($login);
             setcookie('user_name',$data['user_phone'],time()+3600*24);
-            return json_encode(4);
+            return 4;
         }
     }
 
@@ -72,10 +70,9 @@ class IndexController extends Controller
     public function user()
     {
         if(empty($_COOKIE['user_name'])){
-            echo "<script>alert('您好！您还没有登陆');</script>";
-            die;
+            return redirect('home/index');
         }
-        $user = new user();
+        $user = new UserModel();
         $data = $user->selOne($_COOKIE['user_name']);
         return view('home/index/user',['data'=>$data]);
     }
@@ -85,31 +82,29 @@ class IndexController extends Controller
      */
     public function register()
     {
-        session_start();
         $user_phone = Input::get('user_phone');
         $code = Input::get('code');
         $sms = Input::get('sms');
         $user_password = Input::get('user_password');
-        $message = $_SESSION['code'];
-        if($_SESSION['milkcaptcha'] != $code) {
-            return json_encode(0);
+        $message = session('code');
+        if(session('milkcaptcha') != $code) {
+            return 0;
         }
         if($sms != $message) {
-            return json_encode(1);
+            return 1;
         }else{
             $pwd = md5($user_password);
             $data = ['user_phone'=>$user_phone,'user_password'=>$pwd];
-            $user = new user();
+            $user = new UserModel();
             $result = $user->selOne($user_phone);
             if($result) {
-                return json_encode(2);
+                return 2;
             }else {
                 $res = $user->add($data);
                 if($res) {
-                    unset($_SESSION);
-                    session_destroy();
+                    session()->flush();
                     setcookie('user_name',$user_phone,time()+3600*24);
-                    return json_encode(3);
+                    return 3;
                 }
             }
         }
@@ -127,8 +122,7 @@ class IndexController extends Controller
         $phrase = $builder->getPhrase();
 
         //把内容存到session
-        session_start();
-        $_SESSION['milkcaptcha'] = $phrase;
+        session()->put(['milkcaptcha'=>$phrase]);
         ob_clean();//清楚缓存;
         return response($builder->output())->header('content-type','image/jpeg');
     }
@@ -138,41 +132,8 @@ class IndexController extends Controller
      */
     public function loginDo()
     {
-        header("Content-Type:text/html;charset=UTF-8");
         $user_phone = Input::get('user_phone');
         $code = rand(1000,9999);
-        function nowapi_call($a_parm){
-            if(!is_array($a_parm)){
-                return false;
-            }
-            //combinations
-            $a_parm['format']=empty($a_parm['format'])?'json':$a_parm['format'];
-            $apiurl=empty($a_parm['apiurl'])?'http://api.k780.com/?':$a_parm['apiurl'].'/?';
-            unset($a_parm['apiurl']);
-            foreach($a_parm as $k=>$v){
-                $apiurl.=$k.'='.$v.'&';
-            }
-            $apiurl=substr($apiurl,0,-1);
-            if(!$callapi=file_get_contents($apiurl)){
-                return false;
-            }
-            //format
-            if($a_parm['format']=='base64'){
-                $a_cdata=unserialize(base64_decode($callapi));
-            }elseif($a_parm['format']=='json'){
-                if(!$a_cdata=json_decode($callapi,true)){
-                    return false;
-                }
-            }else{
-                return false;
-            }
-            //array
-            if($a_cdata['success']!='1'){
-                echo $a_cdata['msgid'].' '.$a_cdata['msg'];
-                return false;
-            }
-            return $a_cdata['result'];
-        }
         $nowapi_parm['app']='sms.send';
         $nowapi_parm['tempid']='51382';
         $nowapi_parm['param']='code%3D'.$code;
@@ -180,15 +141,46 @@ class IndexController extends Controller
         $nowapi_parm['appkey']='30945';
         $nowapi_parm['sign']='7c17bdef0a7c60575c36f72a9cccc6e3';
         $nowapi_parm['format']='json';
-        nowapi_call($nowapi_parm);
-        session_start();
-        $_SESSION['code']=$code;
+        $this->nowapi_call($nowapi_parm);
+        session()->put(['code'=>$code]);
         return json_encode(1);
     }
 
-    public function cookies()
-    {
-        setcookie("user_name","");
-        return redirect('home/index');
+    /**
+     * @param $a_parm
+     * @return bool
+     * 短信接口
+     */
+    public function nowapi_call($a_parm){
+        if(!is_array($a_parm)){
+            return false;
+        }
+        //combinations
+        $a_parm['format']=empty($a_parm['format'])?'json':$a_parm['format'];
+        $apiurl=empty($a_parm['apiurl'])?'http://api.k780.com/?':$a_parm['apiurl'].'/?';
+        unset($a_parm['apiurl']);
+        foreach($a_parm as $k=>$v){
+            $apiurl.=$k.'='.$v.'&';
+        }
+        $apiurl=substr($apiurl,0,-1);
+        if(!$callapi=file_get_contents($apiurl)){
+            return false;
+        }
+        //format
+        if($a_parm['format']=='base64'){
+            $a_cdata=unserialize(base64_decode($callapi));
+        }elseif($a_parm['format']=='json'){
+            if(!$a_cdata=json_decode($callapi,true)){
+                return false;
+            }
+        }else{
+            return false;
+        }
+        //array
+        if($a_cdata['success']!='1'){
+            echo $a_cdata['msgid'].' '.$a_cdata['msg'];
+            return false;
+        }
+        return $a_cdata['result'];
     }
 }
